requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
django-webpack-loader
pillow
gunicorn
